v5.18.14
Added WIP Bluetooth patch for BCM4377 Macbooks by Sven from Asahilinux.
This will not work if you are using 2.4Ghz WiFi at the same time, or if
you don't have bluetooth firmware installed (get that from macos, and
check in dmesg for where in /lib/firmware/brcm to put it).